Session 2 - Research/Designing.
In this session I dedicated myself to researching how to use GML code to create to of my most crucial mechanics, the camera system and the Enemy AI.

Designs for the buttons, main menu/start up screen have been completed as well.

Note To Self: These designs were made in Game maker and not with any other software.

This is also when I set-up the folders and sprites/Objects in Game maker ready for their designs.

I've also begun in doing the code for the camera system and have made substantial progress in it, this is the foundation of when coding has begun.